requirements.txt
streamlit
scikit-learn
nltk
python-docx
PyMuPDF